Extract endpoint resolving to separate file in workspacerouting
